prediction_model.py
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import MinMaxScaler
from tensorflow.keras.models import Sequential
from tensorflow.keras.layers import LSTM, Dense
from sklearn.ensemble import RandomForestClassifier
from statsmodels.tsa.arima.model import ARIMA
from data_manager import get_historical_data

# Placeholder for feature engineering and data preprocessing
def preprocess_data(data):
    df = pd.DataFrame([(d.issueNumber, d.color, d.size, d.openResult, d.openTime) for d in data],
                      columns=['issueNumber', 'color', 'size', 'openResult', 'openTime'])
    
    # Convert categorical features to numerical
    df['color_encoded'] = df['color'].astype('category').cat.codes
    df['size_encoded'] = df['size'].astype('category').cat.codes
    
    # Example: Create lag features for time-series analysis
    df['openResult_lag1'] = df['openResult'].shift(1)
    df['color_encoded_lag1'] = df['color_encoded'].shift(1)
    
    # Drop rows with NaN values created by shifting
    df.dropna(inplace=True)
    
    return df

# Placeholder for LSTM model
def train_lstm_model(X_train, y_train_color, y_train_size):
    scaler = MinMaxScaler()
    X_train_scaled = scaler.fit_transform(X_train)
    
    model_color = Sequential()
    model_color.add(LSTM(50, activation='relu', input_shape=(X_train_scaled.shape[1], 1)))
    model_color.add(Dense(len(y_train_color.unique()), activation='softmax'))
    model_color.compile(optimizer='adam', loss='sparse_categorical_crossentropy', metrics=['accuracy'])
    
    model_size = Sequential()
    model_size.add(LSTM(50, activation='relu', input_shape=(X_train_scaled.shape[1], 1)))
    model_size.add(Dense(len(y_train_size.unique()), activation='softmax'))
    model_size.compile(optimizer='adam', loss='sparse_categorical_crossentropy', metrics=['accuracy'])
    
    # Reshape for LSTM input [samples, timesteps, features]
    X_train_scaled = X_train_scaled.reshape(X_train_scaled.shape[0], X_train_scaled.shape[1], 1)
    
    model_color.fit(X_train_scaled, y_train_color, epochs=10, batch_size=32, verbose=0)
    model_size.fit(X_train_scaled, y_train_size, epochs=10, batch_size=32, verbose=0)
    
    return model_color, model_size, scaler

# Placeholder for Ensemble model
def train_ensemble_model(X_train, y_train_color, y_train_size):
    model_color = RandomForestClassifier(n_estimators=100, random_state=42)
    model_color.fit(X_train, y_train_color)
    
    model_size = RandomForestClassifier(n_estimators=100, random_state=42)
    model_size.fit(X_train, y_train_size)
    
    return model_color, model_size

# Placeholder for ARIMA model (example for time-series, not directly for classification)
def train_arima_model(data_series):
    # ARIMA is typically for forecasting numerical series, not direct classification
    # This is a placeholder to show its inclusion if a numerical target was present
    model = ARIMA(data_series, order=(5,1,0)) # Example order
    model_fit = model.fit()
    return model_fit

def predict_next_result():
    historical_data = get_historical_data(limit=500)
    if not historical_data:
        return {"error": "No historical data available for prediction."}

    df = preprocess_data(historical_data)
    
    # Define features (X) and targets (y)
    features = ['openResult_lag1', 'color_encoded_lag1'] # Example features
    X = df[features]
    y_color = df['color_encoded']
    y_size = df['size_encoded']
    
    # Split data (for training purposes, in real-time, you'd train on all available data)
    X_train, X_test, y_train_color, y_test_color, y_train_size, y_test_size = train_test_split(
        X, y_color, y_size, test_size=0.2, random_state=42
    )
    
    # Train LSTM models
    lstm_model_color, lstm_model_size, scaler = train_lstm_model(X_train, y_train_color, y_train_size)
    
    # Train Ensemble models
    ensemble_model_color, ensemble_model_size = train_ensemble_model(X_train, y_train_color, y_train_size)
    
    # Run backtesting
    run_backtesting(lstm_model_color, lstm_model_size, scaler, X_test, y_test_color, y_test_size)

    # Get the last available data point for prediction
    last_data_point = df[features].iloc[-1].values.reshape(1, -1)
    
    # Predict with LSTM
    last_data_point_scaled = scaler.transform(last_data_point.reshape(1, -1))
    last_data_point_scaled = last_data_point_scaled.reshape(last_data_point_scaled.shape[0], last_data_point_scaled.shape[1], 1)
    lstm_pred_color_proba = lstm_model_color.predict(last_data_point_scaled)[0]
    lstm_pred_size_proba = lstm_model_size.predict(last_data_point_scaled)[0]
    
    # Predict with Ensemble
    ensemble_pred_color_proba = ensemble_model_color.predict_proba(last_data_point)[0]
    ensemble_pred_size_proba = ensemble_model_size.predict_proba(last_data_point)[0]
    
    # Combine predictions (simple averaging for demonstration)
    combined_pred_color_proba = (lstm_pred_color_proba + ensemble_pred_color_proba) / 2
    combined_pred_size_proba = (lstm_pred_size_proba + ensemble_pred_size_proba) / 2
    
    predicted_color_idx = combined_pred_color_proba.argmax()
    predicted_size_idx = combined_pred_size_proba.argmax()
    
    # Map back to original labels
    color_map = {code: label for label, code in df[['color', 'color_encoded']].drop_duplicates().set_index('color_encoded').to_dict()['color'].items()}
    size_map = {code: label for label, code in df[['size', 'size_encoded']].drop_duplicates().set_index('size_encoded').to_dict()['size'].items()}
    
    predicted_color = color_map.get(predicted_color_idx, "Unknown")
    predicted_size = size_map.get(predicted_size_idx, "Unknown")
    
    confidence_color = combined_pred_color_proba[predicted_color_idx] * 100
    confidence_size = combined_pred_size_proba[predicted_size_idx] * 100
    
    # Placeholder for actual period number (this would come from the latest fetched data)
    # To get the *next* issue number, we need to fetch the latest issue number from the API
    # and increment it. This requires a separate API call or logic to determine the next issue.
    # For now, we'll use a placeholder or a simple increment if the issueNumber is purely numerical.
    # Assuming issueNumber is a string like '20250716100052190', we need to parse and increment the last part.
    # This is a simplification and might need more robust logic based on the actual API behavior.
    try:
        latest_issue_number_str = df["issueNumber"].iloc[-1]
        # Assuming the last few digits are the incrementing part
        prefix = latest_issue_number_str[:-3] # e.g., '202507161000521'
        suffix = int(latest_issue_number_str[-3:]) # e.g., '90'
        next_suffix = str(suffix + 1).zfill(3) # Increment and pad with leading zeros
        next_issue_number = prefix + next_suffix
    except Exception as e:
        print(f"Error determining next issue number: {e}. Using placeholder.")
        next_issue_number = "UNKNOWN_NEXT_PERIOD"

    return {
        "color": predicted_color,
        "size": predicted_size,
        "period_number": next_issue_number, 
        "confidence_score": max(confidence_color, confidence_size)
    }

if __name__ == "__main__":
    # This part needs a running database with some data to work
    # For testing, you might need to manually populate the database or run data_manager.py first
    prediction = predict_next_result()
    print(prediction)




# Placeholder for continuous learning
def update_model_with_new_data(new_data):
    # In a real scenario, this would involve retraining or fine-tuning models
    # with new data. For now, it's a conceptual placeholder.
    print("Model updated with new data (placeholder).")

# Placeholder for backtesting
def run_backtesting(model_color, model_size, scaler, X_test, y_test_color, y_test_size):
    # This function would simulate predictions on historical data and evaluate performance
    # For demonstration, we'll just print a message.
    print("Running backtesting (placeholder).")
    # Example: Evaluate LSTM model
    X_test_scaled = scaler.transform(X_test)
    X_test_scaled = X_test_scaled.reshape(X_test_scaled.shape[0], X_test_scaled.shape[1], 1)
    loss_color, accuracy_color = model_color.evaluate(X_test_scaled, y_test_color, verbose=0)
    loss_size, accuracy_size = model_size.evaluate(X_test_scaled, y_test_size, verbose=0)
    print(f"LSTM Color Model Accuracy: {accuracy_color:.2f}")
    print(f"LSTM Size Model Accuracy: {accuracy_size:.2f}")

    # Example: Evaluate Ensemble model
    ensemble_accuracy_color = model_color.score(X_test, y_test_color)
    ensemble_accuracy_size = model_size.score(X_test, y_test_size)
    print(f"Ensemble Color Model Accuracy: {ensemble_accuracy_color:.2f}")
    print(f"Ensemble Size Model Accuracy: {ensemble_accuracy_size:.2f}")


